Adopts the shared pieces from #63: api.Num for every integer field, the
generic ui.Date, and resolveHackathon from cmd/hackathon_ctx.go in place of
the local helper. Drops host submissions, since hillpost submissions already
lists them, and the local api.Number, now that api.Num does the same job.

+++ cli/cmd/hackathons.go
@@ -87,7 +87,7 @@ var discoverCmd = &cobra.Command{
 		}
 		rows := make([][]string, 0, len(hackathons))
 		for _, h := range hackathons {
-			rows = append(rows, []string{h.Name, ui.Date(int64(h.StartDate)), ui.Date(int64(h.EndDate)), h.ID})
+			rows = append(rows, []string{h.Name, ui.Date(h.StartDate), ui.Date(h.EndDate), h.ID})
 		}
 		fmt.Print(ui.Table([]string{"NAME", "STARTS", "ENDS", "ID"}, rows))
 		fmt.Println("\n" + ui.Label.Render("Join one with: hillpost join --public <id>"))
